First upload
Old code for generating procedural 3d maps and adapted for Dark Basic Pro back in the day. This project did once build with Visual Studio 2008 and produce a dll. I doubt very much that it will now. It's only here as a reminder as to a project I once spent some time on figuring out and relearning trigonometry to get working.